requirements.txt
# Core libraries
numpy==1.26.4
scipy==1.13.1
torch==2.3.1+cpu
torchvision==0.18.1+cpu
opencv-python==4.9.0.74
pillow==12.0.0
networkx==3.6.1
sympy==1.14.0
mpmath==1.3.0
tbb==2021.13.1
intel-openmp==2021.4.0
mkl==2021.4.0

# Web framework
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
aiofiles==23.2.1

# Pydantic v2
pydantic==2.7.1
pydantic-settings==2.3.4

# Visualization / plotting
matplotlib==3.9.0

# XAI
shap==0.45.1

# ML utils
scikit-learn==1.5.0

# Groq free-tier LLM
groq==0.9.0

# Env
python-dotenv==1.0.1

# Optional
filelock==3.20.0
fsspec==2025.12.0
Jinja2==3.1.6
MarkupSafe==3.0.2
typing_extensions==4.15.0